adds checkin screen with fields validation and background integration

+++ Teste-Dev-iOS-Woop-Sicredi/Models/Cupon.swift
@@ -12,5 +12,5 @@ struct Cupon: Decodable {
     
     let id: String
     let eventId: String
-    let discount: Int
+    let discount: Double
 }
